refactor endpoints to individual controllers. Need to fix token validation.
